pyproject.toml
[tool.poetry]
name = "auth-fastapi"
version = "0.1.0"
description = ""
authors = ["cleber <cleber.santos.2099@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi = "^0.103.2"
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"
pydantic = "^2.4.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.23.2"
httpx = "^0.25.0"
pydantic-settings = "^2.0.3"
bcrypt = "^4.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
taskipy = "^1.12.0"
blue = "^0.9.1"
ruff = "^0.1.3"
httpx = "^0.25.0"
isort = "^5.12.0"
freezegun = "^1.2.2"

[tool.ruff]
line-length = 79
exclude = ['.venv', 'alembic']

[tool.isort]
profile = "black"
line_length = 79
extend_skip = ['alembic']

[tool.pytest.ini_options]
pythonpath = "."

[tool.blue]
extend-exclude = '(alembic/)'

[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff'
format = 'blue .  && isort .'
run = 'uvicorn app.main:app --reload'
pre_test = 'task lint'
test = 'pytest -s -x --cov=app -vv'
post_test = 'coverage html'
docker='docker run -d -p 5439:5432 -e POSTGRES_USER=postgres -e POSTGRES_PASSWORD=postgres -e POSTGRES_DB=postgres postgres:latest'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
